+++ packages/serverless-workflow-diagram-editor/lienzo-webapp/src/main/java/org/kie/lienzo/client/AutomaticLayoutExample.java
@@ -192,7 +192,7 @@ public class AutomaticLayoutExample extends BaseExample implements Example {
         }
 
         layer.refresh();
-        layer.draw();
+        layer.batch();
     }
 
     private AbstractLayoutService createLayoutService() {
