+++ packages/serverless-workflow-diagram-editor/sw-editor/sw-editor-client/src/main/java/org/kie/workbench/common/stunner/sw/client/ShapeFactory.java
@@ -19,9 +19,7 @@ package org.kie.workbench.common.stunner.sw.client;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
-import com.ait.lienzo.client.core.shape.Picture;
 import org.appformer.kogito.bridge.client.resource.ResourceContentService;
-import org.appformer.kogito.bridge.client.resource.interop.ResourceContentOptions;
 import org.kie.workbench.common.stunner.core.client.shape.Shape;
 import org.kie.workbench.common.stunner.core.definition.shape.Glyph;
 import org.kie.workbench.common.stunner.sw.client.resources.GlyphFactory;
@@ -73,59 +71,11 @@ public class ShapeFactory
         } else if (instance instanceof End) {
             return new EndShape();
         } else if (instance instanceof State) {
-            State state = (State) instance;
-            StateShape shape = new StateShape(state.getName());
-            if (state.metadata == null) {
-                shape.setType(state.getType());
-                return shape;
-            }
-
-            if (state.metadata.type != null) {
-                shape.setType(state.metadata.type);
-            }
-
-            if (!shape.isIconEmpty()) {
-                return shape;
-            }
-
-            if (state.metadata.icon == null) {
-                shape.setType(state.getType());
-                return shape;
-            }
-
-            if (state.metadata.icon.startsWith("data:")) {
-                Picture picture = new Picture(state.metadata.icon);
-                shape.setIconPicture(picture);
-            } else {
-                resourceContentService
-                        .get(state.metadata.icon, ResourceContentOptions.binary())
-                        .then(image -> {
-                            if (image != null) {
-                                String base64image = iconDataUri(state.metadata.icon, image);
-                                Picture picture = new Picture(base64image);
-                                shape.setIconPicture(picture);
-                            } else {
-                                shape.setType(state.getType());
-                            }
-                            return null;
-                        });
-            }
-
-            return shape;
+            return new StateShape((State) instance, resourceContentService);
         } else if (TransitionShape.isTransition(instance)) {
             return TransitionShape.create(instance).setAppearance(instance);
         }
         return null;
-    }
-
-    protected static String iconDataUri(String iconUri, String iconData) {
-        String[] iconUriParts = iconUri.split("\\.");
-        if (iconUriParts.length > 1) {
-            int fileTypeIndex = iconUriParts.length - 1;
-            String fileType = iconUriParts[fileTypeIndex];
-            return "data:image/" + fileType + ";base64, " + iconData;
-        }
-        return iconData;
     }
 
     @Override
